+++ src/Strava/Strava.API.Client/Model/Athlete/Error/GetAthleteError.cs
@@ -1,0 +1,10 @@
+using Strava.API.Client.Model.Base.Error;
+
+namespace Strava.API.Client.Model.Athlete.Error;
+
+public class GetAthleteError : BaseError
+{
+    public GetAthleteError(string message) : base(message)
+    {
+    }
+}